# apifox-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IM Chat System API",
    "description": "即时通讯系统 RESTful API 文档",
    "version": "1.0.0",
    "contact": {
      "name": "IM Chat Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081/api",
      "description": "本地开发环境"
    }
  ],
  "tags": [
    {
      "name": "用户管理",
      "description": "用户注册、登录、信息管理"
    }
  ],
  "paths": {
    "/user/register": {
      "post": {
        "tags": ["用户管理"],
        "summary": "用户注册",
        "description": "创建新用户账号",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              },
              "examples": {
                "正常注册": {
                  "summary": "正常注册流程",
                  "value": {
                    "username": "testuser001",
                    "password": "123456",
                    "nickname": "测试用户001",
                    "phone": "13800138001",
                    "email": "test001@example.com"
                  }
                },
                "最小必填": {
                  "summary": "只填必填字段",
                  "value": {
                    "username": "testuser002",
                    "password": "123456",
                    "nickname": "测试用户002"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                },
                "examples": {
                  "成功响应": {
                    "summary": "注册成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "userId": 1,
                        "username": "testuser001",
                        "nickname": "测试用户001",
                        "avatar": "https://cube.elemecdn.com/0/88/03b0d39583f48206768a7534e55bcpng.png",
                        "gender": 0,
                        "phone": "13800138001",
                        "email": "test001@example.com",
                        "signature": null
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "用户名已存在": {
                    "value": {
                      "code": 1001,
                      "message": "用户名已存在",
                      "data": null
                    }
                  },
                  "手机号已注册": {
                    "value": {
                      "code": 1005,
                      "message": "手机号已被注册",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["用户管理"],
        "summary": "用户登录",
        "description": "用户账号密码登录",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              },
              "examples": {
                "正常登录": {
                  "summary": "使用正确的账号密码登录",
                  "value": {
                    "username": "testuser001",
                    "password": "123456"
                  }
                },
                "错误密码": {
                  "summary": "密码错误的情况",
                  "value": {
                    "username": "testuser001",
                    "password": "wrongpass"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "examples": {
                  "成功响应": {
                    "summary": "登录成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "token": "eyJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjEsInVzZXJuYW1lIjoidGVzdHVzZXIwMDEiLCJleHAiOjE3MzIwMTIzNDV9.abc123",
                        "userInfo": {
                          "userId": 1,
                          "username": "testuser001",
                          "nickname": "测试用户001",
                          "avatar": "https://cube.elemecdn.com/0/88/03b0d39583f48206768a7534e55bcpng.png",
                          "gender": 0,
                          "phone": "13800138001",
                          "email": "test001@example.com",
                          "signature": null
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "登录失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "用户不存在": {
                    "value": {
                      "code": 1002,
                      "message": "用户不存在",
                      "data": null
                    }
                  },
                  "密码错误": {
                    "value": {
                      "code": 1003,
                      "message": "用户名或密码错误",
                      "data": null
                    }
                  },
                  "账号已禁用": {
                    "value": {
                      "code": 1006,
                      "message": "账号已被禁用",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/info/{id}": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取用户信息",
        "description": "根据用户ID获取用户详细信息",
        "operationId": "getUserInfo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "用户ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "examples": {
              "正常ID": {
                "value": 1
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                },
                "examples": {
                  "成功响应": {
                    "summary": "获取用户信息成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "userId": 1,
                        "username": "testuser001",
                        "nickname": "测试用户001",
                        "avatar": "https://cube.elemecdn.com/0/88/03b0d39583f48206768a7534e55bcpng.png",
                        "gender": 0,
                        "phone": "13800138001",
                        "email": "test001@example.com",
                        "signature": "这是我的个性签名"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "未登录": {
                    "value": {
                      "code": 401,
                      "message": "未认证",
                      "data": null
                    }
                  },
                  "Token过期": {
                    "value": {
                      "code": 1004,
                      "message": "Token已过期",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "用户不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "用户不存在": {
                    "value": {
                      "code": 1002,
                      "message": "用户不存在",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证，格式：Bearer {token}"
      }
    },
    "schemas": {
      "UserRegisterRequest": {
        "type": "object",
        "required": ["username", "password", "nickname"],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名（唯一）",
            "minLength": 3,
            "maxLength": 20,
            "example": "testuser001"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "minLength": 6,
            "maxLength": 20,
            "example": "123456"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "maxLength": 50,
            "example": "测试用户"
          },
          "phone": {
            "type": "string",
            "description": "手机号（可选，唯一）",
            "pattern": "^1[3-9]\\d{9}$",
            "example": "13800138001"
          },
          "email": {
            "type": "string",
            "description": "邮箱（可选）",
            "format": "email",
            "example": "test@example.com"
          }
        }
      },
      "UserLoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "testuser001"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "123456"
          }
        }
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "用户ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "testuser001"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "测试用户"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL",
            "example": "https://cube.elemecdn.com/0/88/03b0d39583f48206768a7534e55bcpng.png"
          },
          "gender": {
            "type": "integer",
            "description": "性别 0-未知 1-男 2-女",
            "example": 0
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138001"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "test@example.com"
          },
          "signature": {
            "type": "string",
            "description": "个性签名",
            "example": "这是我的个性签名"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "description": "JWT Token",
                "example": "eyJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjEsInVzZXJuYW1lIjoidGVzdHVzZXIwMDEiLCJleHAiOjE3MzIwMTIzNDV9.abc123"
              },
              "userInfo": {
                "$ref": "#/components/schemas/UserInfo"
              }
            }
          }
        }
      },
      "UserInfoResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/UserInfo"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误信息",
            "example": "请求参数错误"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "错误详情",
            "example": null
          }
        }
      }
    }
  }
}